+++ apps/bot/src/discord/utils/index.ts
@@ -1,6 +1,7 @@
 export { assertGuildMemberIsAdmin } from './assert-guild-member-is-admin.js';
 export { assertInteractorIsTheOwner } from './assert-interactor-is-the-owner.js';
 export { buildBackAction, buildBackButton } from './build-back-action.js';
+export { buildCancelButton, cancelButtonHandler } from './cancel-button.js';
 export { buildCustomId } from './build-custom-id.js';
 export { buildDiscordTimestamp } from './build-discord-timestamp.js';
 export { buildMenuButtons } from './build-menu-buttons.js';
