+++ apps/bot/src/domains/guild/commands/set-language.ts
@@ -2,8 +2,8 @@ import { SUPPORTED_LANGUAGES } from '@one-piece/db';
 import { ActionRowBuilder, ButtonBuilder, ButtonStyle } from 'discord.js';
 
 import type { Command } from '../../../discord/types.js';
-import { assertGuildMemberIsAdmin, buildOpEmbed, buildCustomId } from '../../../discord/utils/index.js';
-import { CANCEL_SET_LANGUAGE_BUTTON_NAME, LANGUAGE_FLAGS } from '../constants.js';
+import { assertGuildMemberIsAdmin, buildCancelButton, buildOpEmbed, buildCustomId } from '../../../discord/utils/index.js';
+import { LANGUAGE_FLAGS } from '../constants.js';
 
 export const setLanguageCommand: Command = {
   name: 'setlanguage',
@@ -13,10 +13,7 @@ export const setLanguageCommand: Command = {
 
     const embed = buildOpEmbed('info').setTitle('Langue du serveur').setDescription('Dans quelle langue veux-tu que le bot te parle ?');
     const languageRow = new ActionRowBuilder<ButtonBuilder>().addComponents(
-      new ButtonBuilder()
-        .setCustomId(buildCustomId(CANCEL_SET_LANGUAGE_BUTTON_NAME, message.author.id))
-        .setLabel('Annuler')
-        .setStyle(ButtonStyle.Secondary),
+      buildCancelButton(message.author.id),
       ...SUPPORTED_LANGUAGES.map((language) =>
         new ButtonBuilder()
           .setCustomId(buildCustomId(language, message.author.id))
