+++ apps/bot/src/domains/guild/commands/set-prefix.ts
@@ -4,8 +4,8 @@ import { ActionRowBuilder, ButtonBuilder, ButtonStyle } from 'discord.js';
 import { CUSTOM_ID_SEPARATOR } from '../../../discord/constants.js';
 import { ValidationError } from '../../../discord/errors.js';
 import type { Command } from '../../../discord/types.js';
-import { assertGuildMemberIsAdmin, buildCustomId, buildOpEmbed } from '../../../discord/utils/index.js';
-import { CANCEL_SET_PREFIX_BUTTON_NAME, CONFIRM_SET_PREFIX_BUTTON_NAME } from '../constants.js';
+import { assertGuildMemberIsAdmin, buildCancelButton, buildCustomId, buildOpEmbed } from '../../../discord/utils/index.js';
+import { CONFIRM_SET_PREFIX_BUTTON_NAME } from '../constants.js';
 
 const WHITESPACE_REGEX = /\s/;
 
@@ -22,10 +22,7 @@ export const setPrefixCommand: Command = {
       .setDescription(`Es-tu sûr de vouloir changer le préfixe de \`${guild.prefix}\` en \`${prefix}\` ?`);
 
     const row = new ActionRowBuilder<ButtonBuilder>().addComponents(
-      new ButtonBuilder()
-        .setCustomId(buildCustomId(CANCEL_SET_PREFIX_BUTTON_NAME, message.author.id))
-        .setLabel('Annuler')
-        .setStyle(ButtonStyle.Secondary),
+      buildCancelButton(message.author.id),
       new ButtonBuilder()
         .setCustomId(buildCustomId(CONFIRM_SET_PREFIX_BUTTON_NAME, message.author.id, prefix))
         .setLabel('Confirmer')
